+++ forge/changelog.md
@@ -1,1 +1,2 @@
 - Allow creating groups with specific ID using the API
+- Added hidden groups to the API